+++ src/Errors/UniqueValueDulicateException.php
@@ -2,8 +2,7 @@
 
 namespace LBF\Errors;
 
-use Exception;
-use Throwable;
+use LBF\Errors\Meta\ExceptionMeta;
 
 /**
  * Error page for handling errors that occure when a duplicate of a unique value is trying to be set.
@@ -12,24 +11,7 @@ use Throwable;
  * 
  * @author  Gareth Palmer   [Github & Gitlab /projector22]
  * @since   LBF 0.1.11-beta
+ * @since   LBF 0.2.0-beta  Moved __construct to abstract class `ExceptionMeta`.
  */
 
-class UniqueValueDulicateException extends Exception {
-
-    /**
-     * Class constructor.
-     * 
-     * @param   string          $message — [optional] The Exception message to throw.
-     * @param   int             $code — [optional] The Exception code.
-     * @param   Throwable|null  $previous [optional] The previous throwable used for the exception chaining.
-     * 
-     * @return  never
-     * 
-     * @access  public
-     * @since   LBF 0.1.4-beta
-     */
-
-    public function __construct( $message, $code = 0, ?Throwable $previous = null ) {
-        parent::__construct( $message, $code, $previous );
-    }
-}
+class UniqueValueDulicateException extends ExceptionMeta {}